dashboard/requirements.txt
streamlit
pandas
matplotlib
plotly
psycopg2-binary
